requirements.txt
python-docx==0.8.11
markdown==3.5.1
pypandoc==1.11
openai==0.28.0
streamlit==1.28.1
